Added Animation

+++ lib/providers/cart.dart
@@ -23,8 +23,8 @@ class Cart with ChangeNotifier {
     return _items.length;
   }
 
-  double get totalAmount {
-    double total = 0.0;
+  int get totalAmount {
+    int total = 0;
     _items.forEach((key, value) {
       total = total + value.price * value.quantity;
     });
